Save ha labels for ZFS/Group in config
Save ha labels for group containing ZFS RA and Lustre RA names.
This allows "non-standard" HA setups to be controlled via standard
agent calls.
